+++ shardis-ui/src/main/frontend/app/shared/index.ts
@@ -1,2 +1,3 @@
 export * from './x-large'
 export * from './auth'
+export * from './guards'