+++ shardis-ui/src/main/frontend/platform/browser/providers.ts
@@ -2,7 +2,7 @@ import {FORM_PROVIDERS, LocationStrategy, PathLocationStrategy} from '@angular/c
 import {HTTP_PROVIDERS} from '@angular/http';
 import {MATERIAL_PROVIDERS} from './material2';
 import {NG2_WEBSTORAGE} from 'ng2-webstorage';
-import {asyncRoutes, routes} from '../../app/app.routes';
+import {asyncRoutes, routes, AUTH_PROVIDERS} from '../../app/app.routes';
 import {provideWebpack} from './webpack/webpack-component-resolver';
 import {provideRouter} from '@angular/router';
 
@@ -16,6 +16,7 @@ export const APPLICATION_PROVIDERS = [
   ...MATERIAL_PROVIDERS,
   ...NG2_WEBSTORAGE,
   provideRouter(routes),
+  ...AUTH_PROVIDERS,
   provideWebpack(asyncRoutes),
   {
     provide: LocationStrategy,