+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Much cleaner webpack configuration
 * Role based visibility of menu items
 * CRUD example for authenticated users
+* Route access for selected roles (guards)
 
 ### 1.7.0 (08.06.2016)
 * Added npm-check-updates as dev dependency